-fix: invalid names
Changed some code to allow exporting and replacing of file with invalid filenames inside the pak.
Only supported for single file export/replace/add/delete.
Export all will silently fail to create these kinds of files and cannot be used for these.
The file info box now also show the index in the pak FAT.
You can click on the filename in the file info box, to copy it to clipboard.