-add: saving add settings, new warning dialog
You can now save default settings for the Add File dialog.
The warning message when editing now has a option to disable it, and also added a option uncheck the read-only box when opening by default as well.
If you disabled the warning and want to re-enable it, open a pak file while holding shift when you click File -> Open